+++ server/controllers/userController.js
@@ -4,8 +4,7 @@ const db = require('../models')
 const Subscription = db.Subscription
 const User = db.User
 const Category = db.Category
-const crypto = require("crypto"); // 加密
-const { validMessage } = require('../middleware/middleware')
+const { validMessage, getTradeInfo } = require('../middleware/middleware')
 const districts = require('../location/district.json')
 const bcrypt = require('bcryptjs')
 const jwt = require('jsonwebtoken')
@@ -15,96 +14,6 @@ const Op = Sequelize.Op
 
 // import express-validator for validation
 const { validationResult } = require('express-validator/check');
-
-const URL = 'https://f798ad18.ngrok.io'; //本地 domain 不接受，使用 ngrok 工具做臨時網址，取得的網址放這
-const MerchantID = 'MS38035958'; // 商店代號
-const HashKey = 'WF1pmVp4AxMgFs13YrlZQPuirCd47ql6'; //API 金鑰
-const HashIV = 'CFPnopI11rY5YolP'; //API 金鑰
-const PayGateWay = "https://ccore.newebpay.com/MPG/mpg_gateway"; //付款網址
-const ReturnURL = URL + "/api/user/subscribe/spgateway/callback?from=ReturnURL"; //支付完成返還商店網址
-const NotifyURL = URL + "/api/user/subscribe/spgateway/callback?from=NotifyURL"; //支付通知網址
-const ClientBackURL = URL + "/api"; //支付取消返回網址
-
-// 放入 create_mpg_aes_encrypt 將交易資訊轉成字串，以便加密使用
-function genDataChain(TradeInfo) {
-  let results = [];
-  for (let kv of Object.entries(TradeInfo)) {
-    results.push(`${kv[0]}=${kv[1]}`);
-  }
-  return results.join("&");
-}
-
-function create_mpg_aes_encrypt(TradeInfo) {
-  let encrypt = crypto.createCipheriv("aes256", HashKey, HashIV);
-  let enc = encrypt.update(genDataChain(TradeInfo), "utf8", "hex");
-  return enc + encrypt.final("hex");
-}
-
-function create_mpg_sha_encrypt(TradeInfo) {
-  let sha = crypto.createHash("sha256");
-  let plainText = `HashKey=${HashKey}&${TradeInfo}&HashIV=${HashIV}`;
-
-  return sha
-    .update(plainText)
-    .digest("hex")
-    .toUpperCase();
-}
-
-// 交易完成後回傳資料使用的反向解密
-function create_mpg_aes_decrypt(TradeInfo) {
-  let decrypt = crypto.createDecipheriv("aes256", HashKey, HashIV);
-  decrypt.setAutoPadding(false);
-  let text = decrypt.update(TradeInfo, "hex", "utf8");
-  let plainText = text + decrypt.final("utf8");
-  let result = plainText.replace(/[\x00-\x20]+/g, "");
-  return result;
-}
-
-function getTradeInfo(Amt, Desc, email, sn) {
-  console.log("===== getTradeInfo =====");
-  console.log(Amt, Desc, email, sn);
-  console.log("==========");
-
-  data = {
-    MerchantID: MerchantID, // 商店代號
-    RespondType: "JSON", // 回傳格式
-    TimeStamp: Date.now(), // 時間戳記
-    Version: 1.5, // 串接程式版本
-    MerchantOrderNo: (sn) ? sn : Date.now(), // 商店訂單編號
-    LoginType: 0, // 智付通會員
-    OrderComment: "OrderComment", // 商店備註
-    Amt: Amt, // 訂單金額
-    ItemDesc: Desc, // 產品名稱
-    Email: email, // 付款人電子信箱
-    ReturnURL: ReturnURL, // 支付完成返回商店網址
-    NotifyURL: NotifyURL, // 支付通知網址/每期授權結果通知
-    ClientBackURL: ClientBackURL // 支付取消返回商店網址
-  };
-
-  console.log("===== getTradeInfo: data =====");
-  console.log(data);
-
-  mpg_aes_encrypt = create_mpg_aes_encrypt(data);
-  mpg_sha_encrypt = create_mpg_sha_encrypt(mpg_aes_encrypt);
-
-  console.log("===== getTradeInfo: mpg_aes_encrypt, mpg_sha_encrypt =====");
-  console.log(mpg_aes_encrypt);
-  console.log(mpg_sha_encrypt);
-
-  tradeInfo = {
-    MerchantID: MerchantID, // 商店代號
-    TradeInfo: mpg_aes_encrypt, // 加密後參數
-    TradeSha: mpg_sha_encrypt,
-    Version: 1.5, // 串接程式版本
-    PayGateWay: PayGateWay,
-    MerchantOrderNo: data.MerchantOrderNo
-  };
-
-  console.log("===== getTradeInfo: tradeInfo =====");
-  console.log(tradeInfo);
-
-  return tradeInfo;
-}
 
 let userController = {
   emailCheck: async (req, res) => {
@@ -255,14 +164,18 @@ let userController = {
         order: [['createdAt', 'DESC']],
         limit: 1
       })
-
       if (subscription.length > 0 && subscription[0].dataValues.sub_expired_date > Date.now()) {
         return res.status(200).json({ status: 'success', message: 'You still have an active subscrtiption.' })
       }
-      if (subscription.length === 0 ||
-        subscription[0].dataValues.payment_status !== '0' &&
-        subscription[0].dataValues.sub_expired_date < Date.now()
-      ) {
+      // 需要訂購的人有 沒有訂單，有訂單但是過期，有訂單但是餘額用完
+      const newbie = (subscription.length === 0) ? true : false
+      const expired = (subscription[0].dataValues.payment_status !== '0' &&
+        subscription[0].dataValues.sub_expired_date < Date.now()) ? true : false
+      const insufficient = (subscription[0].dataValues.payment_status !== '0' &&
+        subscription[0].dataValues.sub_balance <= 0 &&
+        subscription[0].dataValues.sub_expired_date > Date.now()) ? true : false
+
+      if (newbie || expired || insufficient ) {
         const tradeInfo = getTradeInfo(req.body.sub_price, req.body.sub_name, req.user.email) //req.user.email
         let subscription = await Subscription.create({
           UserId: req.user.id,
@@ -274,9 +187,16 @@ let userController = {
         })
         return res.status(200).json({ status: 'success', subscription, tradeInfo, message: 'Successfully create a subscription' })
       }
-
-      // 如果有訂單
-      const tradeInfo = getTradeInfo(subscription[0].sub_price, subscription[0].sub_name, req.body.email, subscription[0].sn)
+      // 如果有訂單，但還沒付款，先產生新 trandeInfo 記得傳入 sn，如果選擇的方案不相同，修改方案
+      const tradeInfo = getTradeInfo(subscription[0].sub_price, subscription[0].sub_name, req.user.email, subscription[0].sn)
+      if (subscription[0].dataValues.sub_name !== req.body.sub_name) {
+        subscription = await subscription.update({
+          sub_name: req.body.sub_name,
+          sub_price: req.body.sub_price,
+          sub_balance: req.body.sub_balance
+        })
+      }
+      // 如果有訂單，選擇的方案相同
       return res.status(200).json({ status: 'success', subscription, tradeInfo, message: 'you can countinue to describe the NextMeal.' })
     } catch (error) {
       console.log(error)
@@ -292,10 +212,9 @@ let userController = {
       console.log(req.query); // 回傳 { from: NotifyURL}，第四次回傳 { from: ReturnURL}
       console.log(req.body); // 回傳的 object 解碼使用
       console.log("==========");
-
       console.log("===== spgatewayCallback: TradeInfo =====");
       console.log(req.body.TradeInfo);
-
+      // 將回傳交易訊息解密
       const data = JSON.parse(create_mpg_aes_decrypt(req.body.TradeInfo))
       console.log("===== spgatewayCallback: create_mpg_aes_decrypt、data =====");
       console.log(data);
